requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.3
djangorestframework==3.15.1
filelock==3.13.3
pbr==6.0.0
platformdirs==4.2.0
psycopg2==2.9.9
sqlparse==0.4.4
stevedore==5.2.0
tzdata==2024.1
virtualenv==20.25.1
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
virtualenvwrapper-win==1.2.7
